requirements.txt
coverage
flask
mutagen
PyYAML
requests
sqlalchemy
pyjwt

